Fix misuse of sizeof() for string literals
A previous commit replaced occurrences of `strlen( LITERAL )`
by `sizeof( LITERAL )` in order to evaluate the literal length
at compile-time.

This change was flawed since a C string literal includes a
0-terminator, and this 0-Byte is accounted for in sizeof().

While the stack continued to work when tested against itself,
this change made the implementation non-compliant with RFC 8446,
since now various magic TLS 1.3 labels had another 0-Byte at the
end, affecting the key schedule.

This could easily be fixed by replacing each invocation of
`sizeof( LITERAL )` by `sizeof( LITERAL ) - 1`, but this would
still waste an unused 0-Byte at the end of every string literal
which is used as a TLS 1.3 label.

Instead, this commit introduces a globally visible structure
containing all labels as statically sized, non-0-terminated
string literals, and replaces the pattern

```
  LITERAL, sizeof( LITERAL )
```

by

```
  label_structure.label_name, sizeof( label_structure.label_name )
```

The latter is in fact generated by a helper macro

```
   MBEDTLS_SSL_TLS1_3_LBL_WITH_LEN()
```

which only takes the label name as an argument.

One benefit of this approach is that one can infer the maximum
length of all TLS 1.3 labels without the use of awkward nested
`max( max( ... ) )` expressions, nor the use of magic constants,
by placing all labels in a `union` and checking the `sizeof`
this union.

Signed-off-by: Hanno Becker <hanno.becker@arm.com>