Add option to mbedtls_ssl_tls1_3_derive_secret() to hash context
There are a few occasions in TLS 1.3 where an empty message
argument is passed to the `Derive-Secret()` routine. By its definition

```
       Derive-Secret(Secret, Label, Messages) =
            HKDF-Expand-Label(Secret, Label,
                              Transcript-Hash(Messages), Hash.length)
```

this means that we need to hash an empty string.

The previous code hardcoded the hashes of 0-length strings under
SHA-256 and SHA-512 in such cases.

This commit removes such hardcoded constants by adding an option
to `mbedtls_ssl_tls1_3_derive_secret()` -- the function which
implements `Derive-Secret()` in Mbed TLS -- to indicate whether
the context still needs hashing or not.

This allows to pass `NULL, 0, MBEDTLS_SSL_TLS1_3_CONTEXT_UNHASHED`
in the cases where "" shall be used as the context which still needs
to be hashed, and `hash, hash_len, MBEDTLS_SSL_TLS1_3_CONTEXT_HASHED`
in cases where the message argument isn't empty, and the running
transcript hash of the messages is already available.

Signed-off-by: Hanno Becker <hanno.becker@arm.com>