Remove duplication ssl_msg.c vs. ssl_tls13_messaging.c: flush output
This commit removes `mbedtls_ssl_flush_output()` from
`ssl_tls13_messaging.c` and reuses the function of the
same name in `ssl_msg.c`.

While the removed version of `mbedtls_ssl_flush_output()`
was originally used in upstream Mbed TLS, there has been
some crucial change to how the record flushing works: In
the recent version of `mbedtls_ssL_flush_output()`, it is
assumed that the pointer `ssl->out_hdr` points to the _end_
of the (sequence of) record(s) to be flushed, while in the
old version, it was assumed to point to the _beginning_.
(The rationale for this change was the introduction of stacking
multiple records in a datagram, for which it was useful to
always have `ssl->out_hdr` point to where the _next_ record
could be written.) This commit adjusts the version of
`mbedtls_ssl_write_record()` from `ssl_tls13_messaging.c`
(which hasn't yet been consolidated with the upstream
`mbedtls_ssl_write_record()`) accordingly.

Signed-off-by: Hanno Becker <hanno.becker@arm.com>