Remove duplicate of mbedtls_ssl_get_record_expansion()
ssl_tls13_messaging.c and ssl_msg.c contain essentially
the same function mbedtls_ssl_get_record_expansion() which
returns the maximum record expansion.

The version of this function provided in ssl_tls13_messaging.c
is different in that it allows to specify if we're asking for
the expansion of incoming or outgoing records. However, this
functionality doesn't seem to be used anywhere in the library,
and moreover it constitutes an API break compared to upstream
Mbed TLS, which should be avoided if possible.

This commit therefore removes the flexibility of specifying
the direction and re-uses the version of mbedtls_ssl_get_record_expansion()
from ssl_msg.c, removing the variant in ssl_tls13_messaging.c.

Signed-off-by: Hanno Becker <hanno.becker@arm.com>