Remove uses of mbedtls_ssl_get_psk()
`mbedtls_ssl_get_get_psk()` currently blurs two things:
(a) access to the (potentially multiple) PSKs configured
    prior to the handshake
(b) access to the _single_ PSK (potentially `NULL`) that has
    been negotiated for use during the current handshake.
If we offer a PSK but don't use it, this leads to functionally
wrong behavior in the current version of the prototype.

The fact that we currently only allow to configure a single
PSK to offer doesn't help the separation of (a) and (b).

There doesn't seem to be a place for `mbedtls_ssl_get_psk()`
in the TLS 1.3 prototype. Instead, we want to either iterate
over the configured PSKs -- external or resumption/ticket or both,
each potentially multiple times -- or we want to use _the_ specific
PSK that the handshake is going to use.

To better separate (a) and (b), this commit modifies the prototype
to _always_ set `ssl->handshake->psk` to the PSK we're using for the
current handshake, potentially `NULL`.
Then, `mbedtls_ssl_tls1_3_key_schedule_stage_early_data()` can
unconditionally use this PSK to initiate the key schedule.

Further, while fo (a) we will ultimately need an iterator over the
configured PSKs, for now this commit introduces a getter
`mbedtls_ssl_get_psk_to_offer()` which checks if a PSK should be offered,
and if so returns this PSK + its PSK identity. This can be used
by the client when writing the ClientHello.

Signed-off-by: Hanno Becker <hanno.becker@arm.com>