Clarify logic for setting/clearing of handshake PSK
The previous code set the handshake PSK during PSK extension writing,
even if the client doesn't use 0-RTT.

Instead, the handshake PSK should evolve as follows:
1) Initially, it's NULL.
2) If the client uses 0-RTT, it sets the handshake PSK to the first
   offered PSK when preparing to write 0-RTT data.
3) If the client uses 0-RTT, it clears the handshake PSK after
   writing the early data.
4) If the server chooses a PSK in its ServerHello, the client
   sets the handshake PSK to this PSK.

If the client uses 0-RTT and the server accepts it, steps
(3) and (4) could be removed. This, however, can be left for
a later optimization.

This commit adjusts the handshake PSK evolution to follow
the steps (1)-(4) above.

Signed-off-by: Hanno Becker <hanno.becker@arm.com>