Add signature algorithms field to mbedtls_ssl_config
Supported Signature Algorithms structures are different between
TLS1.2 and TLS1.3. Origin `sig_hashes` represents hash field defined
in RFC5246, but TLS1.3 does not define the field. TLS1.3 define an enum
to represent signaure/hash pairs, and the enum is not compatible with
TLS1.2.

So, a new field is needed for supported signature algorithms of TLS1.3.

References:
- [RFC8446 Section 4.2.3](https://datatracker.ietf.org/doc/html/rfc8446#section-4.2.3)
- [RFC5246 Section 7.4.1.4.1](https://datatracker.ietf.org/doc/html/rfc5246#section-7.4.1.4.1)

CustomizedGitHooks: yes
Change-Id: Ibef8e05b1453d2f0fc522eaf5b9ec370a84829e2
Signed-off-by: Jerry Yu <jerry.h.yu@arm.com>