ssl-opt.sh: Remove dependencies on RSA being not enabled
When RSA is enabled the default certificate in tests
is an RSA signed one. Writing of RSA certificates is
not yet supported by the TLS 1.3 prototype, thus to be
able to run TLS 1.3 tests where a certificate writing is
involved even if RSA is enabled, define explicitely to
use ECDSA signed certificates in those tests.

Signed-off-by: Ronald Cron <ronald.cron@arm.com>